Intermediate work
At this point the new details panel is working as I want it to, it still needs to load the details data and I then need to go over the filtering so that it works again (normalizing the data names broke it). CSS also needs a bit of cleaneup and theres probably other stuff.